ai-model/requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-groq
langchain-core
langchain-community
pydantic
python-dotenv
httpx
asyncio